Rework base path.
