+++ .autopr/config.yaml
@@ -1,0 +1,20 @@
+input_providers:
+  - type: git_diff
+
+output_branch_template: autopr/{uuid}
+
+llm_provider:
+  type: openai
+  model: gpt-4
+
+pull_request:
+  title: "✨ AutoPR: {summary}"
+  summary_prompt: |
+    Summarize the purpose of this pull request in 1-2 short sentences.
+  commit_message_prompt: |
+    Write a short, conventional commit message describing the changes in this PR.
+
+triggers:
+  - type: push
+    head_branch_matches: ".*"  # любые ветки
+    run: create_pull_request
