+++ code-factory-processor/src/main/kotlin/com/code/factory/bridge/Bridge.kt
@@ -1,14 +1,13 @@
 package com.code.factory.bridge
 
-import com.code.factory.ApiKeyResolver
 import com.code.factory.coderesolver.CodeResolver
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSDeclaration
-import javax.inject.Inject
+import jakarta.inject.Inject
 
 interface BridgeFactory {
-    fun createMain(): Bridge.BridgeMain
+    fun createMain(apiKey: String): Bridge.BridgeMain
 
     fun createTest(): Bridge.BridgeTest
 
@@ -34,28 +33,15 @@ class BridgeFactoryImpl
     constructor(
         private val codeResolver: CodeResolver,
         private val logger: KSPLogger,
-        private val apiKeyResolveProvider: ApiKeyResolver,
     ) : BridgeFactory {
-        @get:JvmName("getApiKeyProperty")
-        private val apiKey: String by lazy { apiKeyResolveProvider.resolve() ?: error("ApiKey not found.") }
-
-        override fun createMain(): Bridge.BridgeMain =
-            keyLogic(
-                mock = BridgeMainWork(codeResolver, openAiServiceMock(logger)),
-                work = BridgeMainWork(codeResolver, openAiService(apiKey, logger)),
-            )
+        override fun createMain(apiKey: String): Bridge.BridgeMain {
+            return when (apiKey) {
+                "test" -> BridgeMainWork(codeResolver, openAiServiceMock(logger))
+                else -> BridgeMainWork(codeResolver, openAiService(apiKey, logger))
+            }
+        }
 
         override fun createTest(): Bridge.BridgeTest = BridgeTestWork()
 
         override fun createGenerated(): Bridge.BridgeGenerated = BridgeGeneratedWork()
-
-        private fun <B : Bridge> keyLogic(
-            mock: B,
-            work: B,
-        ): B {
-            return when (apiKey) {
-                "test" -> mock
-                else -> work
-            }
-        }
     }
