+++ code-factory-processor/src/main/kotlin/com/code/factory/ksp/SleepKspProcessor.kt
@@ -1,5 +1,6 @@
 package com.code.factory.ksp
 
+import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.processing.SymbolProcessor
 import com.google.devtools.ksp.symbol.KSAnnotated
@@ -9,8 +10,11 @@ interface SleepKspProcessor : SymbolProcessor
 
 class SleepKspProcessorImpl
     @Inject
-    constructor() : SleepKspProcessor {
+    constructor(
+        private val logger: KSPLogger,
+    ) : SleepKspProcessor {
         override fun process(resolver: Resolver): List<KSAnnotated> {
+            logger.warn("Sleep processor sleeps.")
             return emptyList()
         }
     }
