+++ code-factory-processor/src/test/kotlin/KspProcessorTest.kt
@@ -1,93 +1,60 @@
-import com.code.factory.AllDeclarationFinder
-import com.code.factory.CodeFilter
-import com.code.factory.CompileChecker
-import com.code.factory.InterfaceFinder
-import com.code.factory.TestCodeFilter
-import com.code.factory.TestSourcePathResolver
-import com.code.factory.bridge.BridgeFactory
-import com.code.factory.coderesolver.CodeResolver
+import com.code.factory.ApiKeyResolver
+import com.code.factory.ProcessorBindModule
 import com.code.factory.ksp.KspProcessor
-import com.code.factory.ksp.PhaseResolver
-import com.code.factory.ksp.Phases
-import com.google.devtools.ksp.processing.CodeGenerator
+import com.code.factory.ksp.SleepKspProcessor
+import com.code.factory.ksp.WorkActorImpl
 import com.google.devtools.ksp.processing.Resolver
-import com.google.devtools.ksp.symbol.KSDeclaration
+import com.google.devtools.ksp.symbol.KSAnnotated
 import io.kotest.core.spec.style.StringSpec
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.verify
+import io.kotest.matchers.collections.shouldContainExactly
+import io.mockk.*
 
 class KspProcessorTest : StringSpec({
 
-    lateinit var bridgeFactory: BridgeFactory
-    lateinit var kspProcessor: KspProcessor
-    lateinit var codeResolver: CodeResolver
-    lateinit var allDeclarationFinder: AllDeclarationFinder
-    lateinit var interfaceFinder: InterfaceFinder
-    lateinit var resolver: Resolver
-    lateinit var declaration: KSDeclaration
-    lateinit var declarations: Sequence<KSDeclaration>
-    lateinit var compileChecker: CompileChecker
-    lateinit var phaseResolver: PhaseResolver
-    lateinit var codeFilter: CodeFilter
-    lateinit var testCodeFilter: TestCodeFilter
-    lateinit var testSourcePathResolver: TestSourcePathResolver
-    lateinit var codeGenerator: CodeGenerator
+    val workActorFactory = mockk<ProcessorBindModule.WorkActorFactory>()
+    val sleepKspProcessor = mockk<SleepKspProcessor>()
+    val apiKeyResolver = mockk<ApiKeyResolver>()
+    val resolver = mockk<Resolver>(relaxed = true)
 
-    beforeTest {
-        bridgeFactory = mockk(relaxed = true)
-        allDeclarationFinder = mockk()
-        resolver = mockk(relaxed = true)
-        declaration = mockk<KSDeclaration>(relaxed = true)
-        declarations = sequenceOf(declaration)
-        every {
-            allDeclarationFinder.getAllDeclaration(any())
-        } returns declarations
-        codeResolver = mockk(relaxed = true)
-        every {
-            codeResolver.getCodeString(declarations)
-        } returns sequenceOf("code context")
-        interfaceFinder = mockk(relaxed = true)
-        compileChecker = mockk(relaxed = true)
-        phaseResolver = mockk(relaxed = true)
-        codeFilter = mockk(relaxed = true)
-        testCodeFilter = mockk(relaxed = true)
-        testSourcePathResolver = mockk(relaxed = true)
-        codeGenerator = mockk(relaxed = true)
-        kspProcessor =
-            KspProcessor(
-                logger = mockk(relaxed = true),
-                interfaceFinder = interfaceFinder,
-                codeResolver = codeResolver,
-                bridgeFactory = bridgeFactory,
-                compileChecker = compileChecker,
-                phaseResolver = phaseResolver,
-                codeFilter = codeFilter,
-                testCodeFilter = testCodeFilter,
-                testSourcePathResolver = testSourcePathResolver,
-                codeGenerator = codeGenerator,
-            )
+    val workActor = mockk<WorkActorImpl>()
+
+    val kspProcessor =
+        KspProcessor(
+            workActorFactory = workActorFactory,
+            sleepKspProcessor = sleepKspProcessor,
+            apiKeyResolver = apiKeyResolver,
+        )
+
+    "should process using workActor when apiKey is resolved" {
+        val apiKey = "my-api-key"
+        val resultList = listOf(mockk<KSAnnotated>())
+
+        every { apiKeyResolver.resolve(resolver) } returns apiKey
+        every { workActorFactory.create(resolver, apiKey) } returns workActor
+        every { workActor.act() } returns resultList
+
+        val result = kspProcessor.process(resolver)
+
+        result shouldContainExactly resultList
+
+        verify { workActorFactory.create(resolver, apiKey) }
+        verify { workActor.act() }
+        verify { apiKeyResolver.resolve(resolver) }
+        confirmVerified(workActorFactory, workActor, apiKeyResolver, sleepKspProcessor)
     }
 
-    "when phase main should work bridge main" {
-        every {
-            phaseResolver.resolvePhase(any())
-        } returns Phases.Main
-        kspProcessor.process(resolver)
+    "should delegate to sleepKspProcessor when apiKey is not resolved" {
+        val fallbackList = listOf(mockk<KSAnnotated>())
 
-        verify {
-            bridgeFactory.createMain()
-        }
-    }
+        every { apiKeyResolver.resolve(resolver) } returns null
+        every { sleepKspProcessor.process(resolver) } returns fallbackList
 
-    "when phase test should work bridge test" {
-        every {
-            phaseResolver.resolvePhase(any())
-        } returns Phases.Tests
-        kspProcessor.process(resolver)
+        val result = kspProcessor.process(resolver)
 
-        verify {
-            bridgeFactory.createTest()
-        }
+        result shouldContainExactly fallbackList
+
+        verify { sleepKspProcessor.process(resolver) }
+        verify { apiKeyResolver.resolve(resolver) }
+        confirmVerified(workActorFactory, workActor, apiKeyResolver, sleepKspProcessor)
     }
 })
